Improvements for people and register models
